requirements.txt
flask
pymysql
Flask-Session
python-dateutil
